organize some redis utilities
